Expand Left/Right key functionality in file pane
When command line is empty, make Left/Right keys function as Alt-.
(go to parent directory) and Enter respectively; otherwise, retain
their normal function to move the cursor within the command line.
This makes it possible to easily navigate directory structure
without lifting fingers from the cursor key block.

Also, when filter is active and the cursor is in the filter field
at the far right position, Right key act as Enter; some users found
this useful to quickly select one of the filtered files, again with
fingers still on the cursor keys.