execute_cmd(): introduce new argument prompt_user
To give user time to read the output of the command that has just
finished executing, function execute_cmd() prompts user to press
Enter before returning to clex file pane. There are circumstances
(for example, when command is interactive or produces no output)
when we might not want that, so add new boolean argument prompt_user
to function execute_cmd(); if it is set to zero, do not prompt the
user and return immediately, but only if command has finished
successfully.