interactive cmds: Fix the display of error messages
In order to avoid screen flickering when it executes an
interactive command, clex stays on alternate screen and
does not call stop_curses(). But, in this way, the normal
terminal operation is not restored and eventual error
messages intended for primary screen end up garbled and
mixed up with the content of clex file pane.

To fix this, do call stop_curses() (which restores normal
terminal operations, but also switches to primary screen)
immediately followed by issuing smup control sequences to
the terminal, switching it back to secondary screen.