execute: run interactive shell when cmd is ""
When called with an empty command, run an interactive shell instead
in function execute(); the shell executable is run with no options
and no arguments.

Also, no longer declare function execute() as static, and add it to
exec.h, as we will need to call it from the function tty_press_enter
in the following commit.